delete: internal header
