+++ src/tokenize/state/in_normal/by_operator.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "in_normal_internal.h"
+#include "../../tokenize_private.h"
 
 void	by_operator(t_token_store *store, t_token_state *state, char current)
 {
